+++ OrganizationSearch/OrganizationsSearchConfig.js
@@ -9,6 +9,11 @@ const indexesMap = {
   aliases: 'aliases',
   erpCode: 'erpCode',
   taxId: 'taxId',
+  bankingAccountNumber: 'bankingInformation.bankAccountNumber',
+};
+
+const protectedIndexesMap = {
+  [indexesMap.bankingAccountNumber]: 'ui-organizations.banking-information.view',
 };
 
 const indexes = Object.values(indexesMap);
@@ -17,20 +22,47 @@ const placeholderIdsMap = {
   [indexesMap.language]: 'ui-organizations.search.placeholder.language',
 };
 
+const KEYWORD_SEARCH_OPTION = {
+  labelId: 'ui-organizations.search.keyword',
+  value: '',
+};
+
+/**
+ * @deprecated Use getSearchableIndexes(stripes) instead.
+ */
 export const searchableIndexes = [
-  {
-    labelId: 'ui-organizations.search.keyword',
-    value: '',
-  },
-  ...indexes.map(index => ({
+  KEYWORD_SEARCH_OPTION,
+  ...indexes.filter(sIndex => !protectedIndexesMap[sIndex]).map(index => ({
     labelId: `ui-organizations.search.${index}`,
     placeholderId: placeholderIdsMap[index],
     value: index,
   })),
 ];
 
-export const getKeywordQuery = query => indexes.reduce(
+const isSearchableIndexHidden = (stripes, sIndex) => Boolean(
+  protectedIndexesMap[sIndex] && !stripes?.hasPerm(protectedIndexesMap[sIndex]),
+);
+
+export const getSearchableIndexes = (stripes) => [
+  KEYWORD_SEARCH_OPTION,
+  ...indexes.reduce((acc, sIndex) => {
+    if (!isSearchableIndexHidden(stripes, sIndex)) {
+      acc.push({
+        labelId: `ui-organizations.search.${sIndex}`,
+        placeholderId: placeholderIdsMap[sIndex],
+        value: sIndex,
+      });
+    }
+
+    return acc;
+  }, []),
+
+];
+
+export const getKeywordQuery = (query, stripes) => indexes.reduce(
   (acc, sIndex) => {
+    if (isSearchableIndexHidden(stripes, sIndex)) return acc;
+
     if (acc) {
       return `${acc} or ${sIndex}="${query}*"`;
     } else {