+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add the "Is Donor" filter selectable options for filters pane. Refs UIPFO-44.
 * Configure organizations modal to display donor organizations and their `name` and `code` fields only. Refs UIPFO-46.
+* Search organization on bank account number. Refs UIPFO-47.
 
 ## [5.0.0](https://github.com/folio-org/ui-plugin-find-organization/tree/v5.0.0) (2023-10-12)
 [Full Changelog](https://github.com/folio-org/ui-plugin-find-organization/compare/v4.0.0...v5.0.0)